nueva ventana venta

+++ src/clasesjava/Venta.java
@@ -1,7 +1,9 @@
 package clasesjava;
 
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.FloatProperty;
 import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleFloatProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -17,23 +19,23 @@ public class Venta {
     private IntegerProperty numeroCompra;
     private StringProperty nombreArticulo;
     private IntegerProperty cantidadCompra;
-    private FloatProperty precioArticulo;
-    private FloatProperty totalCompra;
+    private DoubleProperty precioArticulo;
+    private DoubleProperty totalCompra;
 
-    public Venta(String nombreArticulo, int cantidadCompra, Float precioArticulo, Float totalCompra) {
+    public Venta(String nombreArticulo, int cantidadCompra, Double precioArticulo, Double totalCompra) {
         this.nombreArticulo = new SimpleStringProperty(nombreArticulo);
         this.cantidadCompra = new SimpleIntegerProperty(cantidadCompra);
-        this.precioArticulo = new SimpleFloatProperty(precioArticulo);
-        this.totalCompra = new SimpleFloatProperty(totalCompra);
+        this.precioArticulo = new SimpleDoubleProperty(precioArticulo);
+        this.totalCompra = new SimpleDoubleProperty(totalCompra);
     }
 
-    public Venta(String numeroFactura, int numeroCompra, String nombreArticulo, int cantidadCompra, Float precioArticulo, Float totalCompra) {
+    public Venta(String numeroFactura, int numeroCompra, String nombreArticulo, int cantidadCompra, Double precioArticulo, Double totalCompra) {
         this.numeroFactura = new SimpleStringProperty(numeroFactura);
         this.numeroCompra = new SimpleIntegerProperty(numeroCompra);
         this.nombreArticulo = new SimpleStringProperty(nombreArticulo);
         this.cantidadCompra = new SimpleIntegerProperty(cantidadCompra);
-        this.precioArticulo = new SimpleFloatProperty(precioArticulo);
-        this.totalCompra = new SimpleFloatProperty(totalCompra);
+        this.precioArticulo = new SimpleDoubleProperty(precioArticulo);
+        this.totalCompra = new SimpleDoubleProperty(totalCompra);
     }
 
     public Venta() {
@@ -72,20 +74,20 @@ public class Venta {
         cantidadCompra = new SimpleIntegerProperty(value);
     }
 
-    public final float getPrecioArticulo() {
+    public final Double getPrecioArticulo() {
         return precioArticulo.get();
     }
 
-    public final void setPrecioArticulo(Float value) {
-        precioArticulo = new SimpleFloatProperty(value);
+    public final void setPrecioArticulo(Double value) {
+        precioArticulo = new SimpleDoubleProperty(value);
     }
 
-    public final float getTotalCompra() {
+    public final Double getTotalCompra() {
         return totalCompra.get();
     }
 
-    public final void setTotalCompra(Float value) {
-        totalCompra = new SimpleFloatProperty(value);
+    public final void setTotalCompra(Double value) {
+        totalCompra = new SimpleDoubleProperty(value);
     }
 
     @Override
